Update dry-component and rename “core” dirs to “component”
