+++ apps/admin/component/admin/application.rb
