+++ apps/admin/component/admin/email.rb
