+++ apps/admin/component/admin/enqueue.rb
