+++ apps/admin/component/admin/page.rb
