+++ apps/admin/component/admin/transactions.rb
