+++ apps/admin/component/admin/view.rb
