+++ apps/main/component/boot.rb
