+++ apps/main/component/main/application.rb
