+++ apps/main/component/main/email.rb
