+++ apps/main/component/main/view.rb
