+++ component/berg/assets.rb
