+++ component/boot.rb
@@ -11,6 +11,6 @@ end
 require_relative "berg/persistence"
 
 app_paths = Pathname(__FILE__).dirname.join("../apps").realpath.join("*")
-Dir[app_paths].each { |f| require "#{f}/core/boot" }
+Dir[app_paths].each { |f| require "#{f}/component/boot" }
 
 require_relative "berg/application"
