+++ component/boot/bugsnag.rb
