+++ component/boot/config.rb
