+++ component/boot/i18n.rb
