+++ component/boot/logger.rb
