+++ component/boot/rom.rb
