+++ spec/app_helper.rb
@@ -8,7 +8,7 @@ require "capybara/poltergeist"
 Dir[SPEC_ROOT.join("support/app/*.rb").to_s].each(&method(:require))
 Dir[SPEC_ROOT.join("shared/app/*.rb").to_s].each(&method(:require))
 
-require SPEC_ROOT.join("../core/boot").realpath
+require SPEC_ROOT.join("../component/boot").realpath
 
 Capybara.app = Berg::Application.app
 Capybara.server_port = 3001
